+++ cmd/commands.go
@@ -1,0 +1,78 @@
+package cmd
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"strings"
+
+	"crhuber/kelp/pkg/config"
+	"crhuber/kelp/pkg/initialize"
+
+	"github.com/spf13/cobra"
+)
+
+func NewBrowseCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "browse",
+		Short: "Run the browse command",
+		Long:  `Run the browse command`,
+		Args: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 1 {
+				return errors.New("usage: kelp browse <project>")
+			}
+			return nil
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			repo := args[0]
+			kc, err := config.FindKelpConfig(repo)
+			if err != nil {
+				fmt.Printf("%s \n", err)
+				os.Exit(1)
+			}
+			config.Browse(kc.Owner, kc.Repo)
+			return nil
+		},
+	}
+}
+
+func NewAddCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "add",
+		Short: "Run the add command",
+		Long:  `Run the add command`,
+		Args: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 1 {
+				return errors.New("usage: kelp add owner/repo")
+			}
+			if len(args) < 2 {
+				args = append(args, "latest")
+			}
+			return nil
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			project := args[0]
+			release := args[0]
+			ownerRepo := strings.Split(project, "/")
+			if len(ownerRepo) < 2 {
+				return fmt.Errorf("use owner/repo format")
+
+			}
+			config.ConfigAdd(ownerRepo[0], ownerRepo[1], release)
+
+			return nil
+		},
+	}
+}
+
+func NewInitCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "init",
+		Short: "Run the init command",
+		Long:  `Run the init command`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			initialize.Initialize()
+			return nil
+		},
+	}
+}
