+++ go.mod
@@ -1,9 +1,10 @@
 module crhuber/kelp
 
-go 1.12
+go 1.16
 
 require (
-	github.com/gabriel-vasile/mimetype v1.0.5
-	github.com/mholt/archiver/v3 v3.3.0
+	github.com/gabriel-vasile/mimetype v1.3.1
+	github.com/mholt/archiver/v3 v3.5.0
+	github.com/spf13/cobra v1.2.1
 	github.com/xi2/xz v0.0.0-20171230120015-48954b6210f8
 )
