+++ pkg/config/config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"encoding/json"
@@ -6,19 +6,30 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
+	"os"
 	"os/exec"
+	"path/filepath"
 	"runtime"
 )
 
-type kelpConfig []kelpPackage
+var githubToken string
+var githubUsername string
+var home, err = os.UserHomeDir()
 
-type kelpPackage struct {
+var KelpDir = filepath.Join(home, "/.kelp/")
+var KelpBin = filepath.Join(home, "/.kelp/bin/")
+var KelpCache = filepath.Join(home, "/.kelp/cache/")
+var KelpConf = filepath.Join(home, "/.kelp/kelp.json")
+
+type KelpConfig []KelpPackage
+
+type KelpPackage struct {
 	Owner   string `json:"Owner"`
 	Repo    string `json:"Repo"`
 	Release string `json:"Release"`
 }
 
-func findKelpConfig(repo string) (kelpPackage, error) {
+func FindKelpConfig(repo string) (KelpPackage, error) {
 	kc := loadKelpConfig()
 	for _, kp := range kc {
 		if kp.Repo == repo {
@@ -26,21 +37,21 @@ func findKelpConfig(repo string) (kelpPackage, error) {
 		}
 	}
 	err := errors.New("package not found in config, try adding it first")
-	kp := kelpPackage{}
+	kp := KelpPackage{}
 	return kp, err
 }
 
-func configAdd(owner, repo, release string) {
-	kp := kelpPackage{
+func ConfigAdd(owner, repo, release string) {
+	kp := KelpPackage{
 		Owner:   owner,
 		Repo:    repo,
 		Release: release,
 	}
 	kp.saveToConfig()
 }
-func loadKelpConfig() kelpConfig {
-	bs, _ := ioutil.ReadFile(kelpConf)
-	var kc kelpConfig
+func loadKelpConfig() KelpConfig {
+	bs, _ := ioutil.ReadFile(KelpConf)
+	var kc KelpConfig
 	err := json.Unmarshal(bs, &kc)
 	if err != nil {
 		fmt.Println(err)
@@ -48,10 +59,10 @@ func loadKelpConfig() kelpConfig {
 	return kc
 }
 
-func (kp kelpPackage) saveToConfig() error {
+func (kp KelpPackage) saveToConfig() error {
 	//kc := loadKelpConfig()
-	bs, _ := ioutil.ReadFile(kelpConf)
-	var kc kelpConfig
+	bs, _ := ioutil.ReadFile(KelpConf)
+	var kc KelpConfig
 	err := json.Unmarshal(bs, &kc)
 
 	if err != nil {
@@ -77,7 +88,7 @@ func (kp kelpPackage) saveToConfig() error {
 			if kp.Repo == c.Repo {
 				c.Release = kp.Release
 				bs, _ := json.MarshalIndent(kc, "", " ")
-				ioutil.WriteFile(kelpConf, bs, 0644)
+				ioutil.WriteFile(KelpConf, bs, 0644)
 				fmt.Println("Config updated!")
 				configUpdated = true
 				break
@@ -87,7 +98,7 @@ func (kp kelpPackage) saveToConfig() error {
 	if !matchFound && !configUpdated {
 		kc = append(kc, kp)
 		bs, _ := json.MarshalIndent(kc, "", " ")
-		ioutil.WriteFile(kelpConf, bs, 0644)
+		ioutil.WriteFile(KelpConf, bs, 0644)
 		fmt.Println("Config added!")
 	}
 
@@ -102,11 +113,11 @@ func list() {
 	}
 }
 
-func inspect() {
+func Inspect() {
 	var err error
 	switch runtime.GOOS {
 	case "darwin":
-		err = exec.Command("open", kelpDir).Start()
+		err = exec.Command("open", KelpDir).Start()
 	default:
 		err = fmt.Errorf("unsupported platform")
 	}
@@ -115,7 +126,7 @@ func inspect() {
 	}
 }
 
-func browse(owner, repo string) {
+func Browse(owner, repo string) {
 	var err error
 	var url string
 	url = fmt.Sprintf("https://github.com/%s/%s/releases", owner, repo)
